corrections added

+++ src/main/java/com/cvut/src/model/explosion/Explosion.java
@@ -14,6 +14,7 @@ import javafx.scene.shape.Rectangle;
 public abstract class Explosion implements GameObject {
     protected Image spriteImage;
     protected Renderparam renderParam;
+    //Present frame
     protected int frame = 0;
 
     //Variable for rendering the image frame
@@ -53,7 +54,7 @@ public abstract class Explosion implements GameObject {
      * @param width frame width
      * @param height frame height
      **/
-    protected abstract void drawSpriteFrame(GraphicsContext graphics, Image spriteImage, double x, double y, int columnNumber, int frameNumber, float width, float height);
+    protected abstract void drawSpriteFrame(GraphicsContext graphics, Image spriteImage, double x, double y, int columnNumber, int frameNumber, double width, double height);
 
     /** Returns Rectangle object
      * @return rectangle of this object
